Add REComposeViewController

+++ sdk/sdk/VKApiPhotos.m
@@ -1,7 +1,7 @@
 //
 //  VKApiPhotos.m
 //
-//  Copyright (c) 2013 VK.com
+//  Copyright (c) 2014 VK.com
 //
 //  Permission is hereby granted, free of charge, to any person obtaining a copy of
 //  this software and associated documentation files (the "Software"), to deal in
@@ -21,13 +21,13 @@
 //  CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 #import "VKApiPhotos.h"
-#import "VKPhotoArray.h"
+#import "VKPhoto.h"
 @implementation VKApiPhotos
-- (VKRequest *)getUploadServer:(int)albumId {
+- (VKRequest *)getUploadServer:(NSInteger)albumId {
 	return [self prepareRequestWithMethodName:@"getUploadServer" andParameters:@{ VK_API_ALBUM_ID : @(albumId) }];
 }
 
-- (VKRequest *)getUploadServer:(int)albumId andGroupId:(int)groupId {
+- (VKRequest *)getUploadServer:(NSInteger)albumId andGroupId:(NSInteger)groupId {
 	return [self prepareRequestWithMethodName:@"getUploadServer" andParameters:@{ VK_API_ALBUM_ID : @(albumId), VK_API_GROUP_ID : @(groupId) }];
 }
 
@@ -35,7 +35,7 @@
 	return [self prepareRequestWithMethodName:@"getWallUploadServer" andParameters:nil];
 }
 
-- (VKRequest *)getWallUploadServer:(int)groupId {
+- (VKRequest *)getWallUploadServer:(NSInteger)groupId {
 	return [self prepareRequestWithMethodName:@"getWallUploadServer" andParameters:@{ VK_API_GROUP_ID : @(groupId) }];
 }
