+++ sdk/sdk/VKUploadPhotoRequest.h
@@ -1,7 +1,7 @@
 //
 //  VKUploadPhotoRequest.h
 //
-//  Copyright (c) 2013 VK.com
+//  Copyright (c) 2014 VK.com
 //
 //  Permission is hereby granted, free of charge, to any person obtaining a copy of
 //  this software and associated documentation files (the "Software"), to deal in
@@ -35,5 +35,5 @@
  @param groupId ID of group (without minus sign) on which wall image should be posted (or nil)
  @return Completed request for executing
  */
-- (instancetype)initWithImage:(UIImage *)image parameters:(VKImageParameters *)parameters albumId:(int)albumId groupId:(int)groupId;
+- (instancetype)initWithImage:(UIImage *)image parameters:(VKImageParameters *)parameters albumId:(NSInteger)albumId groupId:(NSInteger)groupId;
 @end
